+++ src/components/Images/index.ts
@@ -1,3 +1,4 @@
 export * from './Avatar'
+export * from './AvatarUploader'
 export * from './IdentityAvatar'
 export * from './Logo'